requirements.txt
Flask==2.1.2
google-generativeai==1.0.0  # Ensure the correct version
